+++ src/main/java/gg/motd/bukkit/MOTDGGPlugin.java
@@ -2,9 +2,11 @@ package gg.motd.bukkit;
 
 import gg.motd.api.APIClient;
 import gg.motd.api.MOTD;
+import org.bukkit.command.CommandSender;
 import org.bukkit.plugin.java.JavaPlugin;
 import org.bukkit.util.CachedServerIcon;
 
+import java.lang.reflect.Method;
 import java.util.Objects;
 
 public class MOTDGGPlugin extends JavaPlugin {
@@ -27,5 +29,15 @@ public class MOTDGGPlugin extends JavaPlugin {
 
     public APIClient getClient() {
         return client;
+    }
+
+    public boolean hasSpigotMethod() {
+        Class<CommandSender> sender = CommandSender.class;
+        try {
+            Method spigot = sender.getMethod("spigot");
+            return true;
+        } catch (NoSuchMethodException e) {
+            return false;
+        }
     }
 }