+++ CHANGELOG.md
@@ -1,3 +1,2 @@
-The last MOTD id and session are now stored and reused.
-- Running `/motdgg apply` will use the session and id of the last MOTD you applied or uploaded 
-- Running `/motdgg editor` will reuse the same session and id but upload your MOTD and server icon again
+- The session id is no long shown in chat for non-console senders on 1.13+
+- Only show a single confirmation message